refactor: 메시지 생성 네이밍 수정, dto 패키지로 이동
