+++ backend/src/main/java/com/woowacourse/thankoo/coupon/application/CouponService.java
@@ -5,7 +5,7 @@ import com.woowacourse.thankoo.alarm.support.AlarmManager;
 import com.woowacourse.thankoo.common.exception.ErrorType;
 import com.woowacourse.thankoo.coupon.application.dto.CouponRequest;
 import com.woowacourse.thankoo.coupon.domain.CouponContent;
-import com.woowacourse.thankoo.coupon.domain.CouponMessage;
+import com.woowacourse.thankoo.coupon.application.dto.CouponMessage;
 import com.woowacourse.thankoo.coupon.domain.CouponRepository;
 import com.woowacourse.thankoo.coupon.domain.Coupons;
 import com.woowacourse.thankoo.member.domain.Member;
@@ -46,7 +46,7 @@ public class CouponService {
 
     private void sendMessage(final Long senderId, final List<String> emails, final CouponContent couponContent) {
         Member sender = getMember(senderId);
-        AlarmManager.setResources(CouponMessage.create(sender.getName(), emails, couponContent));
+        AlarmManager.setResources(CouponMessage.of(sender.getName(), emails, couponContent));
     }
 
     private Member getMember(final Long senderId) {
