+++ backend/src/main/java/com/woowacourse/thankoo/meeting/application/dto/MeetingMessage.java
@@ -1,4 +1,4 @@
-package com.woowacourse.thankoo.meeting.domain;
+package com.woowacourse.thankoo.meeting.application.dto;
 
 import com.woowacourse.thankoo.alarm.support.Message;
 import com.woowacourse.thankoo.alarm.support.Message.Builder;
@@ -11,7 +11,7 @@ public class MeetingMessage {
 
     public static final String PRETEXT = "\uD83E\uDD70 오늘은 미팅이 있는 날이에요!!";
 
-    public static Message create(final List<String> emails) {
+    public static Message of(final List<String> emails) {
         return new Builder()
                 .email(emails)
                 .title(PRETEXT)
